slackbot/requirements.txt
aiohttp==3.7.4.post0
argcomplete==1.12.3
async-timeout==3.0.1
attrs==21.2.0
boto3==1.18.24
botocore==1.21.24
certifi==2021.5.30
cfn-flip==1.2.3
chardet==4.0.0
charset-normalizer==2.0.4
click==7.1.2
durationpy==0.5
Flask==1.1.4
future==0.18.2
hjson==3.0.2
idna==3.2
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
kappa==0.6.0
MarkupSafe==1.1.1
multidict==5.1.0
pep517==0.11.0
pip-tools==6.2.0
placebo==0.9.0
pyee==7.0.4
python-dateutil==2.8.2
python-slugify==5.0.2
PyYAML==5.4.1
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
slackclient==2.9.3
slackeventsapi==2.2.1
text-unidecode==1.3
toml==0.10.2
tomli==1.2.1
tqdm==4.62.1
#troposphere==3.0.2
troposphere<3
typing-extensions==3.10.0.0
urllib3==1.26.6
Werkzeug==0.16.1
wsgi-request-logger==0.4.6
yarl==1.6.3
zappa==0.53.0
